+++ src/Libs/FShade.GLSL/Assembler.fs
@@ -1512,9 +1512,31 @@ module Assembler =
                     | CType.CFloat(64) -> "lf"
                     | _ -> ""
 
-                let str = v.ToString(System.Globalization.CultureInfo.InvariantCulture)
-                if str.Contains "." || str.Contains "E" || str.Contains "e" then return $"{str}{suffix}"
-                else return $"{str}.0{suffix}"
+                let value =
+                    if isPositiveInfinity v then
+                        match t with
+                        | CType.CFloat 32 -> "uintBitsToFloat(0x7F800000u)"
+                        | CType.CFloat 64 -> "packDouble2x32(uvec2(0x00000000u, 0x7FF00000u))"
+                        | _ -> $"(1.0{suffix} / 0.0{suffix})"
+
+                    elif isNegativeInfinity v then
+                        match t with
+                        | CType.CFloat 32 -> "uintBitsToFloat(0xFF800000u)"
+                        | CType.CFloat 64 -> "packDouble2x32(uvec2(0x00000000u, 0xFFF00000u))"
+                        | _ -> $"(-1.0{suffix} / 0.0{suffix})"
+
+                    elif isNaN v then
+                        match t with
+                        | CType.CFloat 32 -> "uintBitsToFloat(0x7FC00000u)"
+                        | CType.CFloat 64 -> "packDouble2x32(uvec2(0x00000000u, 0x7FF80000u))"
+                        | _ -> $"(0.0{suffix} / 0.0{suffix})"
+
+                    else
+                        let str = v.ToString(System.Globalization.CultureInfo.InvariantCulture)
+                        if str.Contains "." || str.Contains "E" || str.Contains "e" then $"{str}{suffix}"
+                        else $"{str}.0{suffix}"
+
+                return value
 
             | CLiteral.CString v -> return "\"" + v + "\""
         }